Isolate parallel test fixtures
Move the fetch-mocking mutex into the shared test helpers so account and
medium tests serialize their temporary global fetch replacement across
files.

Also make the remote actor fixtures in the timeline and post remote
tests unique per run to avoid cross-test contention on the actor table's
(username, instanceHost) uniqueness.
